Adds better rule validation

+++ src/GnipRules.php
@@ -79,9 +79,14 @@ class GnipRules
      * @param array $rules
      *
      * @return RuleList
+     * @throws InvalidRuleFormatException
      */
     public function delete(array $rules)
     {
+        if (!RuleList::validateRawRuleFormat($rules, false)) {
+            throw new InvalidRuleFormatException();
+        }
+
         $client = new Client();
         $client->delete($this->api_url, [
             'auth'           => $this->auth,
